Manage SMTP username through credstash
So that you won't need to update Elastic Beanstalk configs
in order to cycle the credentials. The way SES's SMTP credentials
work is that they are based on an IAM user's access key pair.
Creating a new credentials means the SMTP user name will also change
in addition to the password.